requirements.txt
numpy==1.20.2
matplotlib==3.4.1
pandas==1.2.4
imageio==2.9.0
imageio-ffmpeg==0.4.3
scipy==1.6.3
scikit-image==0.18.1
scikit-learn==0.24.2
natsort==7.1.1
tqdm==4.60.0
pygeoif==0.7
polarTransform==2.0.0
SimpleITK==2.0.2
openpyxl==3.0.7
pygifsicle==1.0.4   # Requires gifsicle to be installed. Check https://github.com/LucaCappelletti94/pygifsicle for more information.
--find-links https://download.pytorch.org/whl/lts/1.8/torch_lts.html # Required to download 
torch==1.8.2+cu102 
torchvision==0.9.2+cu102
torchinfo==0.1.5
efficientnet-pytorch==0.6.3
segmentation-models-pytorch==0.2.0
ttach==0.0.3 # Required for grad-cam
opencv-python==4.5.3.56 # Required for grad-cam
grad-cam==1.2.9
timm==0.4.12
Pillow==8.3.1
python-pptx==0.6.19